Refine refactor.
