+++ rust/sdk/src/stream/arrow/metadata.rs
